+++ docs/sphinx/changelog.rst
@@ -51,3 +51,4 @@ Internal Changes
 - Added elastic deformation gradient decomposition to linear viscoelasticity (:merge:`20`). By `Nathan Miller`_.
 - Generalized the decomposition of the current elastic deformation gradient to current and previous (:merge:`20`). By `Nathan Miller`_.
 - Added the decomposition of the additional state variable vector into volumetric and isochoric parts (:merge:`20`). By `Nathan Miller`_.
+- Added the computation of the rate multipliers and the integration alpha parameter (:merge:`20`). By `Nathan Miller`_.